+++ plm/linkdb.go
@@ -2,6 +2,7 @@ package plm
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/abates/insteon"
 )
@@ -41,115 +42,73 @@ func (mrr *manageRecordRequest) UnmarshalBinary(buf []byte) error {
 }
 
 type LinkDB struct {
-	plm       *PLM
-	refreshCh chan chan bool
-	linkCh    chan chan []*insteon.Link
-	closeCh   chan chan error
+	plm *PLM
 }
 
 func NewLinkDB(plm *PLM) *LinkDB {
-	db := &LinkDB{
-		plm:       plm,
-		refreshCh: make(chan chan bool),
-		linkCh:    make(chan chan []*insteon.Link),
-		closeCh:   make(chan chan error),
-	}
+	db := &LinkDB{plm}
 
-	go db.readWriteLoop()
 	return db
 }
 
-func (db *LinkDB) readWriteLoop() {
+func (db *LinkDB) Links() ([]*insteon.Link, error) {
 	links := make([]*insteon.Link, 0)
-	var refreshCh chan bool
-	var closeCh chan error
 	rrCh := db.plm.Subscribe([]byte{0x57})
+	defer db.plm.Unsubscribe(rrCh)
 
-loop:
-	for {
-		select {
-		case refreshCh = <-db.refreshCh:
-			insteon.Log.Debugf("Refreshing PLM link database")
-			links = make([]*insteon.Link, 0)
-			resp, err := db.plm.Send(&Packet{Command: CmdGetFirstAllLink})
-			if resp.NAK() {
-				refreshCh <- true
-				close(refreshCh)
-				refreshCh = nil
-			} else if err != nil {
-				insteon.Log.Infof("Error sending GetNextAllLink command: %v", err)
-			}
-		case packet := <-rrCh:
-			// only do something if we are in the process of refreshing
-			if refreshCh != nil {
+	insteon.Log.Debugf("Retrieving PLM link database")
+	resp, err := db.plm.Send(&Packet{Command: CmdGetFirstAllLink})
+	if resp.NAK() {
+		err = ErrNak
+	}
+
+	if err == nil {
+	loop:
+		for {
+			select {
+			case packet := <-rrCh:
 				link := packet.Payload.(*insteon.Link)
 				insteon.Log.Debugf("Received PLM record response %v", link)
 				links = append(links, link)
-				resp, err := db.plm.Send(&Packet{Command: CmdGetNextAllLink})
-				if resp.NAK() {
-					refreshCh <- true
-					close(refreshCh)
-					refreshCh = nil
-				} else if err != nil {
-					insteon.Log.Infof("Error sending GetNextAllLink command: %v", err)
+				var resp *Packet
+				resp, err = db.plm.Send(&Packet{Command: CmdGetNextAllLink})
+				if resp.NAK() || err != nil {
+					break loop
 				}
+			case <-time.After(insteon.Timeout):
+				err = insteon.ErrReadTimeout
+				break loop
 			}
-			// TODO remind me why we would get a record response and not be in refresh mode?
-		case linkCh := <-db.linkCh:
-			insteon.Log.Debugf("Returning PLM link database")
-			newLinks := make([]*insteon.Link, len(links))
-			for i, link := range links {
-				newLink := *link
-				newLinks[i] = &newLink
-			}
-			linkCh <- newLinks
-			close(linkCh)
-		case closeCh = <-db.closeCh:
-			break loop
 		}
 	}
-
-	if refreshCh != nil {
-		close(refreshCh)
-	}
-	closeCh <- nil
-}
-
-func (db *LinkDB) Links() []*insteon.Link {
-	ch := make(chan []*insteon.Link)
-	db.linkCh <- ch
-	return <-ch
-}
-
-func (db *LinkDB) Refresh() error {
-	ch := make(chan bool)
-	db.refreshCh <- ch
-	<-ch
-	return nil
+	return links, err
 }
 
 func (db *LinkDB) RemoveLink(oldLink *insteon.Link) (err error) {
 	numDelete := 0
 	deletedLinks := make([]*insteon.Link, 0)
-	for _, link := range db.Links() {
-		if link.Group == oldLink.Group && link.Address == oldLink.Address {
-			numDelete++
-			if !oldLink.Equal(link) {
-				deletedLinks = append(deletedLinks, link)
+	links, err := db.Links()
+	if err == nil {
+		for _, link := range links {
+			if link.Group == oldLink.Group && link.Address == oldLink.Address {
+				numDelete++
+				if !oldLink.Equal(link) {
+					deletedLinks = append(deletedLinks, link)
+				}
 			}
 		}
-	}
 
-	for i := 0; i < numDelete; i++ {
-		_, err = db.plm.Send(&Packet{Command: CmdManageAllLinkRecord, Payload: &manageRecordRequest{command: LinkCmdDeleteFirst, link: oldLink}})
-		if err != nil {
-			insteon.Log.Infof("Failed to remove link: %v", err)
+		for i := 0; i < numDelete; i++ {
+			_, err = db.plm.Send(&Packet{Command: CmdManageAllLinkRecord, Payload: &manageRecordRequest{command: LinkCmdDeleteFirst, link: oldLink}})
+			if err != nil {
+				insteon.Log.Infof("Failed to remove link: %v", err)
+			}
 		}
-	}
 
-	// add back links that we didn't want deleted
-	for _, link := range deletedLinks {
-		db.AddLink(link)
+		// add back links that we didn't want deleted
+		for _, link := range deletedLinks {
+			db.AddLink(link)
+		}
 	}
 	return err
 }
@@ -171,28 +130,23 @@ func (db *LinkDB) AddLink(newLink *insteon.Link) error {
 }
 
 func (db *LinkDB) Cleanup() (err error) {
-	/*removeable := make([]*insteon.Link, 0)
-	for i, l1 := range db.links {
-		for _, l2 := range db.links[i+1:] {
-			if l1.Equal(l2) {
-				removeable = append(removeable, l2)
+	removeable := make([]*insteon.Link, 0)
+	links, err := db.Links()
+	if err == nil {
+		for i, l1 := range links {
+			for _, l2 := range links[i+1:] {
+				if l1.Equal(l2) {
+					removeable = append(removeable, l2)
+				}
+			}
+		}
+
+		for _, link := range removeable {
+			err = db.RemoveLink(link)
+			if err != nil {
+				break
 			}
 		}
 	}
-
-	for _, link := range removeable {
-		err = db.RemoveLink(link)
-		if err != nil {
-			break
-		}
-	}
-	return err*/
-	return ErrNotImplemented
-}
-
-func (db *LinkDB) Close() error {
-	insteon.Log.Debugf("Closing PLM LinkDB")
-	ch := make(chan error)
-	db.closeCh <- ch
-	return <-ch
+	return err
 }
